move more did doc parsing to common-web
